いろんなところいじったから衝突注意

+++ Force_Star/Assets/User/Sakakibara/Player/PlayScene/PlayerDataBace.cs
@@ -454,7 +454,7 @@ public class PlayerDataBace
     {
         m_airTimer += 1;
     }
-    public void CatchTimers(int addCount)
+    public void CatchTimers(float addCount)
     {
         m_airTimer += addCount;
     }
